Fix no display bug, remove back button
I've removed the back button for now because it doesn't work, and I've
replaced it with a home button. I've also made the links non-ajax, and
put the script tags inside the page block (apparently we're supposed to
do that for jQuery mobile). This seems to have fixed the problem of
JS-generated stuff not appearing on the pages.